requirements.txt
flask==2.0.2
pandas==1.3.4
opencv-python-headless==4.5.4.60
numpy==1.21.4
google-cloud-vision==2.7.1
pdf2image==1.16.0
PyPDF2==1.26.0
pytesseract==0.3.8
Flask-MySQL==1.5.2
flask-bcrypt==1.0.1
openpyxl==3.1.2